+++ y-collections/src/jvmTest/kotlin/com/github/whyrising/y/collections/seq/LazySeqTest.kt
@@ -5,7 +5,7 @@ import com.github.whyrising.y.collections.concretions.list.ChunkedSeq
 import com.github.whyrising.y.collections.concretions.list.PersistentList.Empty
 import com.github.whyrising.y.collections.concretions.list.SeqIterator
 import com.github.whyrising.y.collections.concretions.list.l
-import com.github.whyrising.y.collections.concretions.vector.v
+import com.github.whyrising.y.collections.core.v
 import com.github.whyrising.y.collections.mocks.MockSeq
 import com.github.whyrising.y.collections.util.Murmur3
 import com.github.whyrising.y.utils.runAction